requirements.txt
geopandas
numpy
pandas
pyogrio
pyproj
shapely
openai
matplotlib
scikit-learn
keras
python-dotenv
openpyxl